Split mac environment provider tests
